Reorganize

+++ tests.py
@@ -29,9 +29,6 @@ class TestParsely:
     def setup_method(self, method):
         self.parsely = parsely.Parsely('test_queue', 'test_broker')
 
-    def teardown_method(self, method):
-        parsely._tasks.clear()
-
     def test_task(self):
         self.parsely.task()(task_for_test)
 
@@ -84,53 +81,50 @@ class TestParsely:
                 self.handle(mock_celery_task)
             assert not mock_celery_task.retry.called
 
+    def test_make_producer(self):
+        assert isinstance(self.parsely.make_producer(), falcon.api.API)
+
 
 class TestProducer:
     def setup_method(self, method):
         self.parsely = parsely.Parsely('test_queue', 'test_broker')
         self.parsely.task()(task_for_test)
-        self.producer = parsely.Producer(parsely.HTMLRendler(), True)
+        self.producer = parsely.Producer(
+            parsely.HTMLRendler(), True, 'test_queue', self.parsely._tasks)
         self.mock_req = unittest.mock.MagicMock()
         self.mock_resp = unittest.mock.MagicMock()
         parsely.database.Migrator(parsely.__version__).migrate()
         parsely.database.db.reconnect()
 
     def teardown_method(self, method):
-        parsely._tasks.clear()
         parsely.database.db.get().close()
-        os.remove('test.db')
-
-    def test_undefined_queue(self):
-        with pytest.raises(falcon.HTTPBadRequest) as e:
-            self.producer.on_post(
-                self.mock_req, self.mock_resp, 'undefined_queue', 'undefined_task')
-
-        assert e.value.title == "Undefined queue"
+        if os.path.exists('test.db'):
+            os.remove('test.db')
 
     def test_undefined_task(self):
         with pytest.raises(falcon.HTTPBadRequest) as e:
-            self.producer.on_post(self.mock_req, self.mock_resp, 'test_queue', 'undefined_task')
+            self.producer.on_post(self.mock_req, self.mock_resp, 'undefined_task')
 
         assert e.value.title == "Undefined task"
 
     def test_empty_payload(self):
         self.mock_req.stream.read.return_value = b""
         with pytest.raises(falcon.HTTPBadRequest) as e:
-            self.producer.on_post(self.mock_req, self.mock_resp, 'test_queue', 'task_for_test')
+            self.producer.on_post(self.mock_req, self.mock_resp, 'task_for_test')
 
         assert e.value.title == "Empty payload"
 
     def test_non_json_payload(self):
         self.mock_req.stream.read.return_value = b"This is not a JSON"
         with pytest.raises(falcon.HTTPBadRequest) as e:
-            self.producer.on_post(self.mock_req, self.mock_resp, 'test_queue', 'task_for_test')
+            self.producer.on_post(self.mock_req, self.mock_resp, 'task_for_test')
 
         assert e.value.title == "Payload is not a JSON"
 
     def test_lack_message(self):
         self.mock_req.stream.read.return_value = b"{}"
         with pytest.raises(falcon.HTTPBadRequest) as e:
-            self.producer.on_post(self.mock_req, self.mock_resp, 'test_queue', 'task_for_test')
+            self.producer.on_post(self.mock_req, self.mock_resp, 'task_for_test')
 
         assert e.value.title == "Invalid JSON"
 
@@ -147,7 +141,7 @@ class TestProducer:
         }).encode()
         with pytest.raises(falcon.HTTPBadRequest) as e:
             self.producer.on_post(
-                self.mock_req, self.mock_resp, 'test_queue', 'task_for_preprocessor_test')
+                self.mock_req, self.mock_resp, 'task_for_preprocessor_test')
 
         assert e.value.title == "Missing required filed"
 
@@ -166,7 +160,7 @@ class TestProducer:
         }).encode()
         with pytest.raises(falcon.HTTPBadRequest) as e:
             self.producer.on_post(
-                self.mock_req, self.mock_resp, 'test_queue', 'task_for_preprocessor_test')
+                self.mock_req, self.mock_resp, 'task_for_preprocessor_test')
 
         assert e.value.title == "Invalid type"
 
@@ -187,12 +181,12 @@ class TestProducer:
         }).encode()
 
         self.producer.on_post(
-            self.mock_req, self.mock_resp, 'test_queue', 'task_for_preprocessor_test')
+            self.mock_req, self.mock_resp, 'task_for_preprocessor_test')
 
         assert self.parsely._tasks['task_for_preprocessor_test'][0][0].apply_async.called
 
     def test_on_get(self):
-        self.producer.on_get(self.mock_req, self.mock_resp, 'test_queue', 'task_for_test')
+        self.producer.on_get(self.mock_req, self.mock_resp, 'task_for_test')
         self.mock_resp.content_type = 'text/html'
 
 
@@ -222,9 +216,9 @@ class TestStaticResource:
 
 class TestMigrator:
     def teardown_method(self, method):
-        parsely._tasks.clear()
         parsely.database.db.get().close()
-        os.remove('test.db')
+        if os.path.exists('test.db'):
+            os.remove('test.db')
 
     def test__raise_for_invalid_version(self):
         parsely.database.Migrator(parsely.__version__).migrate()
@@ -284,52 +278,13 @@ class TestDBManager:
 
 class TestListTaskResource:
     def setup_method(self, method):
-        self.resource = parsely.TaskListResource(parsely.HTMLRendler())
-
-    def teardown_method(self, method):
-        parsely._tasks.clear()
+        self.resource = parsely.TaskListResource(parsely.HTMLRendler(), 'test', {})
 
     def test_on_get(self):
-        queue_name = 'test_queue'
-        parsely.Parsely(queue_name, 'test_broker')
         mock_resp = unittest.mock.MagicMock()
-        self.resource.on_get(unittest.mock.MagicMock(), mock_resp, queue_name)
+        self.resource.on_get(unittest.mock.MagicMock(), mock_resp)
 
         assert mock_resp.content_type == 'text/html'
-
-    def test_on_get_unknown_queue(self):
-        mock_resp = unittest.mock.MagicMock()
-        with pytest.raises(falcon.HTTPNotFound):
-            self.resource.on_get(unittest.mock.MagicMock(), mock_resp, "unknow_queue")
-
-    def test__list_task_name(self):
-        queue_name = 'test_queue'
-        b = parsely.Parsely(queue_name, 'test_broker')
-        assert len(list(self.resource._list_task_name(queue_name))) == 0
-
-        b.task()(task_for_test)
-        assert len(list(self.resource._list_task_name(queue_name))) == 1
-
-
-class TestQueueListResource:
-    def setup_method(self, method):
-        self.rendler = unittest.mock.MagicMock()
-        self.resource = parsely.QueueListResource(self.rendler)
-
-    def teardown_method(self, method):
-        parsely._tasks.clear()
-
-    def test_on_get(self):
-        queue_names = ["first_quque", "second_queue"]
-        for queue_name in queue_names:
-            parsely.Parsely(queue_name, 'test_broker')
-
-        self.resource.on_get(unittest.mock.MagicMock(), unittest.mock.MagicMock())
-        self.rendler.render.assert_called_once_with("queue_list.html", queue_names=queue_names)
-
-
-def test_producer():
-    assert isinstance(parsely.producer(), falcon.api.API)
 
 
 def test_fibonacci_wait():
